# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API pública de Negociación",
    "version": "1.0.0",
    "description": "Documentación automática generada con swagger-jsdoc"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/usuario/register": {
      "post": {
        "summary": "Registrar un nuevo usuario",
        "tags": [
          "Usuarios"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "nombre",
                  "email",
                  "password"
                ],
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado correctamente"
          }
        }
      }
    },
    "/api/usuario/login": {
      "post": {
        "summary": "Iniciar sesión de usuario",
        "tags": [
          "Usuarios"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesión iniciada correctamente"
          }
        }
      }
    },
...
    "/api/usuario/{id}": {
      "get": {
        "summary": "Obtener un usuario por ID",
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID del usuario"
          }
        ],
        "responses": {
          "200": {
            "description": "Datos del usuario"
          }
        }
      }
    },
    "/api/usuario/update/{id}": {
      "put": {
        "summary": "Actualizar un usuario",
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID del usuario a actualizar"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuario actualizado correctamente"
          }
        }
      }
    },
    "/api/usuario/delete/{id}": {
      "delete": {
        "summary": "Eliminar un usuario",
        "tags": [
          "Usuarios"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "integer"
            },
            "required": true,
            "description": "ID del usuario a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario eliminado correctamente"
          }
        }
      }
    }
  },
...
}